requirements.txt
# requirements.txt

# Core packages
opencv-python==4.3.0.36
pytesseract==0.3.7
numpy==1.18.5
tensorflow==2.3.1
Pillow==7.2.0
matplotlib==3.3.0
tensorflow-gpu==2.3.1
imutils==0.5.3
scikit-learn==0.23.1
pymongo==3.11.3

# Additional packages
tensorflow-datasets==4.4.0
tensorflow-gpu-estimator==2.3.0
scikit-fuzzy==0.4.2
scikit-image==0.18.1
tqdm==4.55.0
tesseract-ocr==4.1.1
tk==8.6.10
pillow==7.2.0
pandas==1.2.0
Pillow==7.2.0
matplotlib==3.3.0
numpy==1.18.5
tensorflow==2.3.1
scikit-learn==0.23.1
pytesseract==0.3.7
matplotlib==3.3.0
pymongo==3.11.3
opencv-python==4.3.0.36
tensorflow-gpu==2.3.1
imutils==0.5.3
Pillow==7.2.0
